new minigame pattern start
