--- plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/minigames/PrototypingMinigame.kt
@@ -1,6 +1,0 @@
-package dev.betrix.superSmashMobsBrawl.minigames
-
-import dev.betrix.superSmashMobsBrawl.models.MinigamePlayer
-
-class PrototypingMinigame(minigameId: String, gameId: String, players: List<MinigamePlayer>) :
-    FfaMinigame(minigameId, gameId, players)
