+++ plugin/src/main/kotlin/dev/betrix/superSmashMobsBrawl/minigames/PrototypingMinigameOld.kt
@@ -1,0 +1,6 @@
+package dev.betrix.superSmashMobsBrawl.minigames
+
+import dev.betrix.superSmashMobsBrawl.models.MinigamePlayer
+
+class PrototypingMinigameOld(minigameId: String, gameId: String, players: List<MinigamePlayer>) :
+    FfaMinigameOld(minigameId, gameId, players)
